requirements.txt
streamlit
transformers
torch
sentencepiece
PyMuPDF
python-docx